adds hacky bugfix for zero length path
When running the app from my office in the computer center and checking
a path to the Engineering Building, the application was crashing.  For
some reason it was getting a zero length path, even though it is
resolving the path to EB correctly.  This hack just fixes the crashing
problem, but we need to figure out what the actual cause of the zero
length path at that stage in the code is.